Reverted back to previous randomisation
Tried a different technique for randomising the list. I tried updating
the players list, however for the sake of future functionality, it is
better to keep the original lsit in order and store match data in a new
list. This means the app does not have to be hard refreshed in order to
start a second tournament with that player list.